#6986 recorder & player multiple fixes

+++ acris-recorder/acris-recorder-client-core/src/main/java/sk/seges/acris/recorder/client/event/fields/IEventFields.java
@@ -2,9 +2,7 @@ package sk.seges.acris.recorder.client.event.fields;
 
 public interface IEventFields {
 	
-	//boolean isStringMapper();
-
 	FieldDefinition getFieldDefinition();
 
-	int getValue();
+	long getValue();
 }
